statstest: split handletest
